Fix log prob calculation

+++ whisper_s2t/backends/ctranslate2/model.py
@@ -256,7 +256,7 @@ class WhisperModelCT2(WhisperModel):
         group_timestamps = []
         for i, segment in enumerate(result):
             # Calculate log probabilities from logits
-            logits = torch.tensor(np.array(segment.logits))
+            logits = torch.tensor(np.array(segment.logits, dtype=np.float32))
             probs = torch.nn.functional.softmax(logits, dim=-1)
             log_probs = torch.log(probs)
             print("log_probs:", log_probs)
